+++ src/PhoneFork.App/Views/MediaView.xaml.cs
@@ -1,0 +1,8 @@
+using System.Windows.Controls;
+
+namespace PhoneFork.App.Views;
+
+public partial class MediaView : UserControl
+{
+    public MediaView() => InitializeComponent();
+}